Follows stylelint naming conventions, adds test
